# test/openapi/openapi.snapshot.json
{
...
  "info": {
    "description": "Deterministic billing simulation and invoice operations API.",
    "title": "LedgerFlow API",
    "version": "0.2.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/customers/{customerId}/billing-profile": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "customerId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/simulations/{runId}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "runId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/v1/customers/{customerId}/billing-profile": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "customerId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/v1/simulations/{runId}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "runId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  }
}